Improved integration test coverage for users, session, lobbies, and players

+++ db/Stored Procedures/dbo.Players_Remove.sql
@@ -1,13 +1,36 @@
-CREATE PROCEDURE [dbo].[Players_Remove] 
+CREATE PROCEDURE [dbo].[Players_Remove]
 	@PlayerId INT
 AS
 BEGIN
 	SET NOCOUNT ON;
 
-	IF NOT EXISTS(SELECT 1 FROM Players WHERE PlayerId = @PlayerId)
+	DECLARE @UserID INT = (SELECT TOP 1 UserId FROM Players WHERE PlayerId = @PlayerId)
+	DECLARE @PlayerTypeId TINYINT = (SELECT TOP 1 PlayerTypeId FROM Players WHERE PlayerId = @PlayerId)
+	DECLARE @LobbyId INT = (SELECT TOP 1 LobbyId FROM Players WHERE PlayerId = @PlayerId)
+	DECLARE @LobbyCreatorUserId INT = (SELECT TOP 1 CreatedByUserId FROM Lobbies WHERE LobbyId = @LobbyId)
+
+	IF @PlayerTypeId IS NULL
 		THROW 50404, 'Player not found.', 1
-                        
-    DELETE FROM Players
-    WHERE PlayerId = @PlayerId
-	
+
+	BEGIN TRAN
+		--If the removed player is the user who created the lobby, close the lobby
+		IF (@UserId IS NOT NULL)
+		AND (@PlayerTypeId = 1) --USER
+		AND (@UserId = @LobbyCreatorUserId)
+		BEGIN
+			DELETE FROM Players WHERE LobbyId = @LobbyId
+			DELETE FROM Lobbies WHERE LobbyId = @LobbyId
+		END
+
+		ELSE
+		BEGIN
+			--If the removed player is a user, remove that user and all guests
+		    IF @PlayerTypeId = 1 --USER
+				DELETE FROM Players WHERE UserId = @UserId
+
+            --Otherwise just delete that player
+			ELSE
+				DELETE FROM Players WHERE PlayerId = @PlayerId
+		END
+	COMMIT
 END
